Various enhancements and fixes to format_fractional_part_nsecs().
Have format_fractional_part_nsecs() return the number of bytes
formatted.  This matches what snprintf() and strlcpy() do, for the
benefit of routines constructing a string in a fixed-length buffer with
multiple calls, one of which is format_fractional_part_nsecs().

Have format_fractional_part_nsecs() check for a count of nanoseconds
greatre than or equal to 1 billion, and report that specially.

Fix places wher it used the bufer pointer and size of the buffer, rather
than the current position in the buffer and the remaining space in the
buffer, to append to the buffer and make sure doing so doesn't overrun
the buffer.

Rename some variable names (rename buf_remaining to just remaining,
rename num_chars to num_bytes as it counts bytes).

Fix some comments.