Add support for the new timestamp precisions to NetMon.
Add WTAP_TSPREC_ values for the new precisions, and rename the 1/10
second and 1/100 second values to match the style used for the new ones,
with additional WTAP_TSPREC_DSEC and WTAP_TSPREC_CSEC defines for
backwards compatibility.  Switch to those new ones in existing files.

For NetMon, use WTAP_TSPREC_100_NSEC for the 2.3 version of the file
format and WTAP_TSPREC_USEC for previous 2.x versions.  Get rid of code
that would have handled 3.x and later formats, as 1) those formats were
never created, 2) the code that was there wouldn't have worked if the
format were different enough, and 3) we didn't support them when reading
in any case.

Regenerate the introspection enums.