Use INT_MIN instead of -2147483648.
Fun fact: -2147483648 isn't an "integer-constant" in C, it's an
expression that applies the unary minus operator to the integer-constant -
2147483648.  The Visual Studio C compiler, at least, treats 2147483648
as an *unsigned* integer constant and warns about applying the unary
minus operator to it.

What we want is the smallest possible int value (which can't be validly
negated on two's-complement platforms), so just use INT_MIN and rely on
the compiler+header supplier to figure out how to define INT_MIN so as
not to provoke a compiler warning.  (At least some versions of Visual
Studio define it as (-2147483647 - 1).)

Fixes #19340.