pcapng: improve handling of time stamp resolutions.
For base-10 resolutions, handle all resolutions from seconds to
nanoseconds.  Reject exponents > 19, as the scale factor won't fit in a
64-bit integer (and even an exponent of 19 means that the largest
possible time stamp can only handle a capture started at midnight UTC,
January 1, 1970, and running for less than 2 seconds, so an exponent
that large is useless in pcapng anyway).

For base-2 resolutions, reject exponents > 63, as the scal efactor won't
fit in a 64-bit integer (and even an exponent of 19 means that the
largest possible time stamp can only handle a capture started at
midnight UTC, January 1, 1970, and running for less than 2 seconds, so
an exponent that large is useless in pcapng anyway).