Prevent fvalue leaks
Add a cleanup function in proto_tree_new_item for the newly
created field_info's fvalue_t, so that if we throw an exception
before adding the new node to the tree, the fvalue_t is properly
freed. Pop it off without calling it if we do add the node.

This is the function that adds nodes from values retrieved from
the tvbuffer, so it can throw exceptions, particularly on variable
length items like FT_STRINGZ, VARINTs, and FT_UINT_[STRING|BYTES].