Enable UTF-8 validation with ENABLE_DEBUG
It is annoying to have to manage two CMake flags independently
for debugging.

Instead of simply removing ENABLE_DEBUG_UTF_8 make it enabled by
default with ENABLE_DEBUG. So ENABLE_DEBUG enables everything and
ENABLE_DEBUG_UTF_8=Off can be used to disable only UTF-8 validation,
if that is desirable for some reason.