Qt: Fix width of "show up to..." prefs in some styles
For some styles, e.g. adwaita, the style adds extra space
around the contents. When setting the maximum width of the
widget, this has to be taken into account, so ask the style
how wide we need to make the line edits to fit our contents.
Otherwise, it's not wide enough to fit the text.