Qt+Windows: Allow updating with an open file with no unsaved changes
Add a FileCloseContext for updates, and use that when calling
testCaptureFileClose in softwareUpdateRequested.

In that context, rather than opening another dialog, just return
false and try again later if there are unsaved changes, but go
ahead and close the file if there are no changes (or if the
user has set the preference not to be asked about unsaved changes.)

This allows updating if the user has a file open with no changes,
e.g. when opening Wireshark by double-clicking an existing file.

Note that even though the WinSparkle dialogs aren't modal, and
a user can bring the main window to the front, save or close the
file and then click on Install Update again without having to
redownload the update, it seems like a lot of users don't expect
that. They also don't know about clicking on Help->Check for Updates
in the menu bar, and think that the automatic update is the only
way to check for updates.

Fix #17658