mgcp, sdp: disable handling SDP RTP if Osmux is announced in MGCP body
"X-Osmux: (*|<cid>)" is an Osmocom extension to MGCP which allows
to signal request and acceptance of use of Osmux framing instead of RTP.
It usually comes together with a regular SDP section which defines the
AMR data to be transported (in Osmux frames rather than RTP packets).
Hence, if Osmux is signalled, we want to avoid telling the RTP subsystem
to decode the recorded packets as RTP, which would then override the
default Osmux dissector.

This is accomplished by passing a setup_info struct to SDP dissector
from MGCP, similar to what's already done in MEGACO and SIP protocols.
An extra bool field is introduced to signal whether Osmux packets are
expected instead of RTP.