HTTP: File data is bytes, not a string
The file data field for HTTP is the Content. It is a stream of
octets, and doesn't have a defined encoding (unlike headers)
inside the HTTP dissector; encodings can be interpreted by
the appropriate media handle or other subdissector.

Don't treat the content as a string with ASCII encoding; make it
an FT_BYTES, like the UDP and TCP payload fields. Some content
types are text and might have an encoding, but others are simply
binary data.

Related to #19280