merge: Allow unlimited number of files
merging is embarrassingly parallel, so long as we preserve ordering
when merging intermediate results, so we can process files in
batches if we hit RLIMIT_NOFILE and have the same output[1] we would
otherwise.

[1] One minor exception is in the SHB comments, but currently those aren't
written properly anyway. Another minor exception is that the ordering
of addresses within the NRBs can vary (be reversed), but both of these
are underlying issues, not having to do with the algorithm of batch
processing and merging.

Wireshark currently supports merging lots of files at once through
drag and drop, but not through the File->Merge dialog; we might
want to change the latter.

Fix #17598