Qt: Store Show Packet Bytes settings in recent
Store the Show Packt Bytes "Decode As" and "Show As" settings in
recent.

The options for "Show as" for Show Packet Bytes are a superset of the
"Show data as" for Follow Stream, and they're conceptually doing
the same thing. Use the same enum for both - Follow Stream doesn't
support all these options, but maybe it should in the future.

To do this, we break the strict correspondance between the index
in the respective comboboxes and the value of the enum. Use
findData when setting the current index so that this doesn't matter;
this also keeps us from having to be as strict about keeping new
entries in alphabetical order.

Fix #17796