dfilter: Add support for BASE_CUSTOM value strings
BASE_CUSTOM is not a value string so it doesn not have a catalog
of values but for implementation purposes we can treat it as a
value string that cannot be uniquely mapped to an integer.

This way fields using BASE_CUSTOM can also be matched using the
custom string representation.

This is probably less useful thani normal value strings but still
it is a use case we can support with the current compiler.