lemon: Add -- support to end options, use in our lemon
Patch lemon to stop parsing options upon encountering "--".
This prevents an issue in the unlikely situation where the
source directory or a source filename has a '=' in it, or
starts with a "-".

Redo 3716933046a565c06da3b637c91cd7f1af36e23b but only
use the -- option in our lemon, not when using system lemon.