[#19362] afp: on't pass null pinfo
When I created !12290, I didn't realize that the `pinfo` parameter to
the function `decode_name_label()` served a dual purpose: if passed as
null, the function skipped changing a column. I assumed it would always
be non-null and that caused #19362.

I've added a separate parameter to `decode_name_label()` to control the
column changing, and pass `pinfo` to all calls. The output structure
should be unchanged, and now this function can use `pinfo->pool` to
allocate string memory.