wsutil: Optimize memmem some more
When the needle is length 1, just use memchr
When the needle is longer, use memchr to check the first byte
before calling memcmp on the rest, so we can skip through the loop
faster.

The GNU memmem returns the haystack when the needle length is 0,
which matches the standard C behavior for strstr when the needle
string is null. Do that, so that our portable version matches the
GNU version on systems that have it. (This shouldn't cause a problem,
because we've being doing that on systems with GNU glibc already.)