DRBD: Add support for dissecting the DRBD "lb-tcp" transport
Since DRBD 9.2.6, a load-balancing transport is available for DRBD. This
is called "lb-tcp". Add support for dissecting this format.

This transport is a wrapper around the plain DRBD format. Each lb-tcp
packet contains part of this DRBD stream. The DRBD packets may be broken
across multiple lb-tcp packets. Conversely, one lb-tcp packet may
contain multiple DRBD packets.

One lb-tcp connection consists of multiple TCP streams. We cannot rely
on having all of these streams in the capture. Even when we do, it is
difficult to associate them with each other. Hence we decode each TCP
stream separately and accept that we may not be able to dissect every
DRBD packet fully.

In practice, most lb-tcp payloads do start on a DRBD packet boundary, so
the result is still very usable.