SSH: Set up channel pairings properly
The channel numbers are not necessarily the same on either side.
Process the SSH_MSG_CHANNEL_OPEN_CONFIRMATION to set up channel
pairings. Previously it was creating a channel pairing when
a subdissector was set but always using channel 0 for both sides.

The message that sets the channel mapping (CONFIRMATION) is from the
opposite side from messages that configure the channel (CHANNEL_OPEN,
CHANNEL_REQUEST), so try to handle this.

We can set a subdissector to one side even if we don't get the
above message (a one sided capture?)

Change a few unsigned ints to uint32_t (as specified in RFC 4254),
and ensure that strings are retrieved in UTF-8 (per RFC 4254).

Note that we still don't handle closed and reassigned channel numbers
for random packet access, which requires something like a multimap
instead of maps.

Related to #16054