wsutil: Do not use built in strcasestr
System implementations of strcasestr are locale dependent; even
though we always force to a UTF-8 locale, the Turkic languages
produce a different result, as the upper case of i is U+0130,
Latin Capital Letter I With Dot Above, and the lower case of I
is is U+0131, Latin Small Letter Dotless i.

For the purposes of dissectors we want the locale-independent ASCII
search. (There may be exceptions in some user-facing GUI code.)

FreeBSD, NetBSD, and macOS (but not OpenBSD currently) provide
strcasestr_l, which could be used with the NULL (C) locale to
produce the desired result.