gcrypt: Disable FIPS mode if possible
libgcrypt 1.10 added a new control command to force disable FIPS
mode, regardless of any file or environment variable that would turn
it on.

FIPS mode makes it impossible to _decrypt_ non-NIST approved
algorithms, both old insecure ones as well as simply unapproved
ones like ChaCha20 and Poly1305. We're decrypting, not promising
security.

Related to #18441 (but doesn't help with distributions like the
one in that bug, that have gcrypt older than 1.10.0)