Frame: dissector table for PcapNG custom blocks
The `frame` dissector has a case statement for dealing with PcapNG
[custom blocks][1], and has some hard-coding in for specific types of
custom block, currently TCP Black Box Log (bblog) and pcaplog blocks.

Propose replacing case statement with a new sub-dissector table keyed on
the Private Enterprise Number, tentatively named `pcapng_custom_block`.

Factor out bblog code from `packet-frame.c` into `packet-bblog.c`. Have
it register itself with this new table.

Create `packet-pcaplog.c` and factor out pcaplog code to that file. Have
it register itself with this new table.

This is one possible way of addressing the concern discussed in !3990 of
generalizing the handling of custom blocks. The hope is to reduce the
tightness of coupling `packet-frame.c` with every new custom block we
want to support.

There remains, unfortunately, tight coupling with custom block
*[options][2]* that exist on standard PcapNG blocks, since the current
established precedent is to give fields for supported options a `frame.`
prefix and put them in the Frame protocol tree.  This proposal treats
that as out of scope.

There is also bblog-specific code in `wiretap/pcapng.c`. It may be
possible to move that into `packet-bblog.c`. This is also beyond the
scope of this proposal.

[1]: https://www.ietf.org/archive/id/draft-tuexen-opsawg-pcapng-05.html#name-custom-block
[2]: https://www.ietf.org/archive/id/draft-tuexen-opsawg-pcapng-05.html#name-custom-options