merge: Add merge comment
The merge code tries to add a comment about what file names
are part of the merge. However, it's been doing so by appending
that comment information to the first comment in the first SHB
of the first source file, and doing nothing if there isn't an
existing comment.

Ensure that the comment is always there by adding the comment
instead. Perhaps it should be prepended instead.

Fix #13839