Qt: Edit capture comments in separate dialog
The capture file properties dialog already lists the section
comments in the Details section, so it wastes some space to have
them a second time in the edit box.

Have an edit dialog for the capture comments. Have the edit dialog
use tabs so that we can edit multiple comments in a section, and
all sections. Allow adding new comments, removing comments, and
reordering comments.

Fix a few small leaks.

Related to #14599, #16133.