dfilter: Improve type inferrence for constant values
The display filter language is statically typed, but the syntax
is not declarative. This means every literal type must be
inferred at compile time. This is not trivial if we have a constant
expression on the left-hand side.

To improve the flexibility and generality of the compiler we add a
look-ahead type resolution step when the type of a value cannot be
immediately inferred.

This simplifies the code, improves some error messages and enables the
correct handling of some expressions that were not supported before,
such as "1 == frame.number".