dfilter: Add time multiplication by scalar integers
The multiplication of two time values is not well-defined,
because time is represented internally as a vector.

Add a scalar multiplication/division for time values
using integer numbers. The scalar multiplier must appear
on the RHS of the operation. (This limitation mat be
removed in the future.)

This is useful to compare relative time values. The
operation is also allowed for absolute date and time values
because it is mathematically consistent but that is
probably less useful in practice.

Related to #19150.