Qt: Improve FollowStreamText scroll performance
We turn on mouse tracking in FollowStreamText so that mouse
moves without any buttons pressed will cause the hint text
to update showing to what packet it will jump if clicked.

However, we don't want to send the mouse moves with no buttons
pressed to the base QPlainTextEdit class, which causes a large
number of calculations for no reason that slow down scrolling
especially in large documents.

Ignore those events, effectively turning off mouse tracking for
the base QPlainTextEdit.