Qt: Fix resetting follow tap when the filter changes
If a Follow Stream is in progress and the tap is reset, e.g.
because the main filter is cleared or another dialog causes a
retap, make sure that the tap is reset.

Prevents the follow information (both displayed and saved) from
having duplicate results, because the old half-performed tap
payload wasn't being cleared.