Qt: Follow Stream stream and substream spinboxes keyboard handling
Don't disable the stream and substream spinboxes when a tap is
in progress, but set them to readonly. When disabled, they can't
have focus, which causes focus to shift to a button when changing
the stream id via keyboard.

Ensure that the spinboxes have keyboard tracking off, so that
valueChanged is only emitted when Enter is pressed or focus changes,
instead of with each number input. (This may differ by default on
different platforms and themes, so ensure it's off.)

Fix #18264