Use explicit test for Clang with compiler attributes
Clang likes to pretend to be other compilers but does not implement
the full range of pragmas other compilers have. Make Clang
always use __attribute__, it should be better supported than
the alternatives, for example with clang-cl.

Do not use __GNUC__ to select Clang, it does not work with
clang-cl. Test explicitly for __clang__ instead.

Fixes #19385.