prefs: GUI preferences don't affect dissection
The GUI preferences don't affect dissection in general and shouldn't
require a costly redissection.

We used to have a PREF_EFFECT_GUI, but it seems that nothing needed
to check it because any changes were noticed in other ways, so it
was removed. (d95213afb0b78f98757d188ef041730fc4a4e3e8) Bring it
back, because the Lua set_preference assumes that *some* effect
flag is returned when a preference is changed.

We have to set this separately for each submodule too, because
the effect flags are reset to PREF_EFFECT_DISSECTION regardless
of the parent module. The changes to columns, fonts, and colors
are also handled separately without special effect flags; the
layout submodule does have its own effect flag.

Fix #17629