radiotap: Improve indication of 6 GHz band channels
Update the channel information string for the 6 GHz band.
Use "6 GHz"; replace "BG" and "A" with "2.4 GHz" and "5 GHz"
as the current labels are somewhat anachronistic now?

Add a comment about 802.11ax certified devices and control frames,
which are transmitted in legacy mode using the basic rate of
6 Mb/s (MCS 0); with no HE or EHT header, these are shown as
PHY_11A, which may be confusing to users.)

Fix #17393